Refactor scripts to update flash fees and add levers
