+++ scripts/governance/add/addLever/addStEthLever/addStethLever.sh
@@ -3,10 +3,10 @@
 
 set -eux
 export HERE=$(dirname $0)
-export CONF=$PWD/$HERE/stEthLever.config.ts
+export CONF=$PWD/$HERE/addStethLever.config.ts
 RUN="npx hardhat run --network tenderly"
 
 $RUN $HERE/../../../../shared/deploy.ts
-$RUN $HERE/stEthLeverSetup.ts
+$RUN $HERE/addStethLever.ts
 $RUN $HERE/../../../../shared/approve.ts
 $RUN $HERE/../../../../shared/execute.ts
